Unified omniscript library code into 2 libraries
