+++ benchmarks/__init__.py
@@ -5,8 +5,12 @@ from pathlib import Path
 
 def create_benchmark(bench_name):
     try:
-        return importlib.import_module(f'.{bench_name}', __name__).Benchmark()
-    except ModuleNotFoundError:
+        return importlib.import_module(f".{bench_name}", __name__).Benchmark()
+    except ModuleNotFoundError as f:
+        # Check if missing module was not our current module, then it's a problem of another module
+        if f != f"No module named '{bench_name}'":
+            # Passthrough this error
+            raise
         available_benchmarks = [p.name for p in Path(__name__).iterdir() if p.is_dir()]
         raise ValueError(
             f'Attempted to create benchmark "{bench_name}", but it is missing from '
