+++ benchmarks/hm_fashion_recs/hm_utils.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 
 import numpy as np
-from utils.pandas_backend import pd
+
+from omniscripts.pandas_backend import pd
 
 
 def load_data(preprocessed_data_path):
