+++ benchmarks/hm_fashion_recs/tm.py
@@ -1,4 +1,4 @@
-from utils import TimerManager
+from omniscripts import TimerManager
 
 # Allow overwrite for now, because notebook is currently calling timer many times
 # TODO: set to false when accurate benchmarking of each iteration is installed
