--- environment/__init__.py
@@ -1,1 +1,0 @@
-from .environment import CondaEnvironment
