+++ omniscripts/benchmark_utils.py
@@ -1,23 +1,22 @@
+"""Utils to be used by inividual benchmarks"""
 import glob
 import os
-import time
-import warnings
 from timeit import default_timer as timer
 
 import psutil
 
-from report import BenchmarkDb
-from benchmarks import create_benchmark
-from utils_base_env import DbConfig
-
-from .namespace_utils import import_pandas_into_module_namespace
-from .pandas_backend import set_backend
-from .benchmark import BaseBenchmark, BenchmarkResults
+from .modin_utils import import_pandas_into_module_namespace
 
 
 repository_root_directory = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 directories = {"repository_root": repository_root_directory}
 
+__all__ = ['files_names_from_pattern', 'load_data_pandas', 'load_data_modin_on_hdk',
+    'split',
+    'print_results',
+    'memory_usage',
+    'getsize'
+    ]
 
 def load_data_pandas(
     filename,
@@ -186,16 +185,6 @@ def split(X, y, test_size=0.1, stratify=None, random_state=None, optimizer="inte
     return (X_train, y_train, X_test, y_test), split_time
 
 
-def check_support(current_params, unsupported_params):
-    ignored_params = {}
-    for param in unsupported_params:
-        if current_params.get(param) is not None:
-            ignored_params[param] = current_params[param]
-
-    if ignored_params:
-        warnings.warn(f"Parameters {ignored_params} are ignored", RuntimeWarning)
-
-
 def memory_usage():
     process = psutil.Process(os.getpid())
     return process.memory_info().rss / (1024**3)  # GB units
@@ -211,105 +200,3 @@ def getsize(filename: str):
         raise ValueError(f"bad s3like link: {filename}")
     else:
         return os.path.getsize(filename) / 1024 / 1024
-
-
-def run_benchmarks(
-    bench_name: str,
-    data_file: str,
-    dfiles_num: int = None,
-    iterations: int = 1,
-    validation: bool = False,
-    optimizer: str = None,
-    pandas_mode: str = "Pandas",
-    ray_tmpdir: str = "/tmp",
-    ray_memory: int = 200 * 1024 * 1024 * 1024,
-    no_ml: bool = None,
-    use_modin_xgb: bool = False,
-    gpu_memory: int = None,
-    extended_functionality: bool = False,
-    db_config: DbConfig = None,
-    commit_hdk: str = "1234567890123456789012345678901234567890",
-    commit_omniscripts: str = "1234567890123456789012345678901234567890",
-    commit_modin: str = "1234567890123456789012345678901234567890",
-):
-    """
-    Run benchmarks for Modin perf testing and report results.
-
-    Parameters
-    ----------
-    bench_name : str
-        Benchmark name.
-    data_file : str
-        A datafile that should be loaded.
-    dfiles_num : int, optional
-        Number of datafiles to load into database for processing.
-    iterations : int, default: 1
-        Number of iterations to run every query. The best result is selected.
-    validation : bool, default: False
-        Validate queries results (by comparison with Pandas queries results).
-    optimizer : str, optional
-        Optimizer to use.
-    pandas_mode : str, default: "Pandas"
-        Specifies which version of Pandas to use: plain Pandas, Modin runing on Ray or on Dask or on HDK.
-    ray_tmpdir : str, default: "/tmp"
-        Location where to keep Ray plasma store. It should have enough space to keep `ray_memory`.
-    ray_memory : int, default: 200 * 1024 * 1024 * 1024
-        Size of memory to allocate for Ray plasma store.
-    no_ml : bool, optional
-        Do not run machine learning benchmark, only ETL part.
-    use_modin_xgb : bool, default: False
-        Whether to use Modin XGBoost for ML part, relevant for Plasticc benchmark only.
-    gpu_memory : int, optional
-        Specify the memory of your gpu(This controls the lines to be used. Also work for CPU version).
-    extended_functionality : bool, default: False
-        Extends functionality of H2O benchmark by adding 'chk' functions and verbose local reporting of results.
-    db_config: DbConfig, optional
-        Configuration for the database
-    commit_hdk : str, default: "1234567890123456789012345678901234567890"
-        HDK commit hash used for benchmark.
-    commit_omniscripts : str, default: "1234567890123456789012345678901234567890"
-        Omniscripts commit hash used for benchmark.
-    commit_modin : str, default: "1234567890123456789012345678901234567890"
-        Modin commit hash used for benchmark.
-    """
-
-    # Set current backend, !!!needs to be run before benchmark import!!!
-    set_backend(pandas_mode=pandas_mode, ray_tmpdir=ray_tmpdir, ray_memory=ray_memory)
-
-    benchmark: BaseBenchmark = create_benchmark(bench_name)
-
-    run_parameters = {
-        "data_file": data_file,
-        "dfiles_num": dfiles_num,
-        "no_ml": no_ml,
-        "use_modin_xgb": use_modin_xgb,
-        "optimizer": optimizer,
-        "pandas_mode": pandas_mode,
-        "ray_tmpdir": ray_tmpdir,
-        "ray_memory": ray_memory,
-        "gpu_memory": gpu_memory,
-        "validation": validation,
-        "extended_functionality": extended_functionality,
-        "commit_hdk": commit_hdk,
-        "commit_omniscripts": commit_omniscripts,
-        "commit_modin": commit_modin,
-    }
-
-    run_id = int(round(time.time()))
-    print(run_parameters)
-
-    reporter = db_config and BenchmarkDb(db_config.create_engine())
-
-    for iter_num in range(1, iterations + 1):
-        print(f"Iteration #{iter_num}")
-        results: BenchmarkResults = benchmark.run(run_parameters)
-
-        if reporter is not None:
-            reporter.report(
-                iteration_no=iter_num,
-                name2time=results.measurements,
-                params=results.params,
-                benchmark=bench_name,
-                run_id=run_id,
-                run_params=run_parameters,
-            )
