+++ omniscripts/timer.py
