--- utils/__init__.py
@@ -1,14 +1,0 @@
-from .utils import (
-    check_support,
-    files_names_from_pattern,
-    import_pandas_into_module_namespace,
-    load_data_pandas,
-    load_data_modin_on_hdk,
-    split,
-    print_results,
-    memory_usage,
-    getsize,
-    run_benchmarks,
-)
-from .timer import TimerManager
-from .benchmark import BaseBenchmark, BenchmarkResults
